+++ extensions/flags/js/package.json
@@ -6,7 +6,7 @@
     "devDependencies": {
         "@types/mithril": "^2.0.8",
         "prettier": "^2.5.1",
-        "flarum-webpack-config": "^2.0.0",
+        "flarum-webpack-config": "^3.0.0",
         "webpack": "^5.76.0",
         "webpack-cli": "^4.9.1",
         "@flarum/prettier-config": "^1.0.0",
