+++ extensions/markdown/js/package.json
@@ -4,7 +4,7 @@
     "version": "0.0.0",
     "prettier": "@flarum/prettier-config",
     "dependencies": {
-        "flarum-webpack-config": "^2.0.0",
+        "flarum-webpack-config": "^3.0.0",
         "webpack": "^5.76.0",
         "webpack-cli": "^4.9.1"
     },
@@ -19,7 +19,7 @@
         "@flarum/prettier-config": "^1.0.0",
         "flarum-tsconfig": "^1.0.2",
         "prettier": "^2.5.1",
-        "flarum-webpack-config": "^2.0.0",
+        "flarum-webpack-config": "^3.0.0",
         "webpack": "^5.76.0",
         "webpack-cli": "^4.9.1"
     }
