+++ extensions/mentions/js/package.json
@@ -5,7 +5,7 @@
     "prettier": "@flarum/prettier-config",
     "devDependencies": {
         "prettier": "^2.5.1",
-        "flarum-webpack-config": "^2.0.0",
+        "flarum-webpack-config": "^3.0.0",
         "webpack": "^5.76.0",
         "webpack-cli": "^4.9.1",
         "@flarum/prettier-config": "^1.0.0"
