+++ extensions/pusher/js/package.json
@@ -19,7 +19,7 @@
         "@flarum/prettier-config": "^1.0.0",
         "@types/pusher-js": "^5.1.0",
         "flarum-tsconfig": "^1.0.2",
-        "flarum-webpack-config": "^2.0.0",
+        "flarum-webpack-config": "^3.0.0",
         "prettier": "^2.5.1",
         "typescript": "^4.5.4",
         "typescript-coverage-report": "^0.6.1",
