+++ extensions/statistics/js/package.json
@@ -10,7 +10,7 @@
         "@flarum/prettier-config": "^1.0.0",
         "@types/mithril": "^2.0.11",
         "flarum-tsconfig": "^1.0.2",
-        "flarum-webpack-config": "^2.0.0",
+        "flarum-webpack-config": "^3.0.0",
         "prettier": "^2.7.1",
         "typescript": "^4.7.4",
         "typescript-coverage-report": "^0.6.4",
