+++ extensions/subscriptions/js/package.json
@@ -5,7 +5,7 @@
     "prettier": "@flarum/prettier-config",
     "devDependencies": {
         "@flarum/prettier-config": "^1.0.0",
-        "flarum-webpack-config": "^2.0.0",
+        "flarum-webpack-config": "^3.0.0",
         "prettier": "^2.7.1",
         "webpack": "^5.76.0",
         "webpack-cli": "^4.10.0",
