+++ extensions/suspend/js/package.json
@@ -13,7 +13,7 @@
     "devDependencies": {
         "@flarum/prettier-config": "^1.0.0",
         "flarum-tsconfig": "^1.0.0",
-        "flarum-webpack-config": "^2.0.0",
+        "flarum-webpack-config": "^3.0.0",
         "prettier": "^2.5.1",
         "webpack": "^5.76.0",
         "webpack-cli": "^4.9.1"
