+++ extensions/tags/src/Api/DiscussionResourceFields.php
@@ -37,7 +37,7 @@ class DiscussionResourceFields
             Schema\Relationship\ToMany::make('tags')
                 ->includable()
                 ->writable()
-                ->required(fn (Context $context, Discussion $discussion) => ! $context->getActor()->can('bypassTagCounts', $discussion))
+                ->required(fn (Context $context, Discussion $discussion) => $context->creating() && ! $context->getActor()->can('bypassTagCounts', $discussion))
                 ->set(function (Discussion $discussion, array $newTags, Context $context) {
                     $actor = $context->getActor();
