+++ framework/core/js/package.json
@@ -34,7 +34,7 @@
         "cross-env": "^7.0.3",
         "expose-loader": "^3.1.0",
         "flarum-tsconfig": "^1.0.2",
-        "flarum-webpack-config": "^2.0.0",
+        "flarum-webpack-config": "^3.0.0",
         "prettier": "^2.5.1",
         "typescript": "^4.5.4",
         "typescript-coverage-report": "^0.6.1",
