+++ framework/core/js/src/admin/components/AdvancedPage.tsx
@@ -4,7 +4,8 @@ import type { IPageAttrs } from '../../common/components/Page';
 import type Mithril from 'mithril';
 import Form from '../../common/components/Form';
 import extractText from '../../common/utils/extractText';
-import FormSectionGroup, { FormSection } from './FormSectionGroup';
+import FormSectionGroup from './FormSectionGroup';
+import FormSection from './FormSection';
 import ItemList from '../../common/utils/ItemList';
 import InfoTile from '../../common/components/InfoTile';
 import { MaintenanceMode } from '../../common/Application';
