+++ framework/core/less/common/Select.less
@@ -11,7 +11,7 @@
     cursor: pointer;
     line-height: 1;
 
-    .FormSection & {
+    .FormSection &, &.FormControl {
       width: 100%;
     }
   }
