+++ framework/core/less/forum/Post.less
@@ -34,8 +34,8 @@
 
   .UserCard {
     position: absolute;
-    top: 15px;
-    left: 16px;
+    top: 0;
+    left: 0;
     z-index: var(--zindex-dropdown);
     transition: opacity 0.2s, transform 0.2s;
     transform: scale(0.95);
@@ -454,7 +454,6 @@
 
   .Post-avatar {
     .Avatar--size(32px);
-    display: block;
   }
 
   .Post-main {
